Move MyDenseVectorUDT to org.apache.spark package because of UserDefinedType is in private spark package for Spark 3.1.x
